Create record_batch_model.py
